requirements.txt
streamlit
matplotlib
seaborn
geopandas